+++ youlai-mall-service/yshop-pms/src/main/java/com/fly4j/yshop/pms/controller/admin/PmsBrandController.java
@@ -61,7 +61,7 @@ public class PmsBrandController extends BaseController {
     }
 
     @ApiOperation(value = "品牌详情", httpMethod = "GET")
-    @ApiImplicitParam(name = "id", value = "品牌id", required = true, paramType = "path", dataType = "Long")
+    @ApiImplicitParam(name = "id", value = "品牌id", required = true, paramType = "path", dataType = "Integer")
     @GetMapping("/{id}")
     public Result get(@PathVariable Long id) {
         PmsBrand brand = iPmsBrandService.getById(id);
@@ -70,7 +70,7 @@ public class PmsBrandController extends BaseController {
 
     @ApiOperation(value = "修改品牌", httpMethod = "PUT")
     @ApiImplicitParams({
-            @ApiImplicitParam(name = "id", value = "品牌id", required = true, paramType = "path", dataType = "Long"),
+            @ApiImplicitParam(name = "id", value = "品牌id", required = true, paramType = "path", dataType = "Integer"),
             @ApiImplicitParam(name = "brand", value = "实体JSON对象", required = true, paramType = "body", dataType = "PmsBrand")
     })
     @PutMapping(value = "/{id}")
@@ -81,7 +81,7 @@ public class PmsBrandController extends BaseController {
     }
 
     @ApiOperation(value = "删除品牌", httpMethod = "DELETE")
-    @ApiImplicitParam(name = "ids", value = "id集合", required = true, paramType = "query", allowMultiple = true, dataType = "Long")
+    @ApiImplicitParam(name = "ids", value = "id集合", required = true, paramType = "query", allowMultiple = true, dataType = "Integer")
     @DeleteMapping()
     public Result delete(@RequestParam("ids") List<Long> ids) {
         boolean status = iPmsBrandService.removeByIds(ids);
