+++ youlai-mall-service/yshop-sms/src/main/java/com/fly4j/yshop/sms/controller/admin/SmsSeckillPromotionController.java
@@ -65,7 +65,7 @@ public class SmsSeckillPromotionController extends BaseController {
 
     @ApiOperation(value = "秒杀活动详情", httpMethod = "GET")
     @ApiImplicitParams({
-            @ApiImplicitParam(name = "id", value = "秒杀活动id", required = true, paramType = "path", dataType = "Long"),
+            @ApiImplicitParam(name = "id", value = "秒杀活动id", required = true, paramType = "path", dataType = "Integer"),
     })
     @GetMapping("/{id}")
     public R get(@PathVariable Long id) {
@@ -75,7 +75,7 @@ public class SmsSeckillPromotionController extends BaseController {
 
     @ApiOperation(value = "修改秒杀活动", httpMethod = "PUT")
     @ApiImplicitParams({
-            @ApiImplicitParam(name = "id", value = "秒杀活动id", required = true, paramType = "path", dataType = "Long"),
+            @ApiImplicitParam(name = "id", value = "秒杀活动id", required = true, paramType = "path", dataType = "Integer"),
             @ApiImplicitParam(name = "smsSeckillPromotion", value = "实体JSON对象", required = true, paramType = "body", dataType = "SmsSeckillPromotion")
     })
     @PutMapping(value = "/{id}")
@@ -85,7 +85,7 @@ public class SmsSeckillPromotionController extends BaseController {
     }
 
     @ApiOperation(value = "删除秒杀活动", httpMethod = "DELETE")
-    @ApiImplicitParam(name = "ids", value = "秒杀活动id", required = true, paramType = "query", allowMultiple = true, dataType = "Long")
+    @ApiImplicitParam(name = "ids", value = "秒杀活动id", required = true, paramType = "query", allowMultiple = true, dataType = "Integer")
     @DeleteMapping()
     public R delete(@RequestParam("ids") List<Long> ids) {
         boolean status = iSmsSeckillPromotionService.removeByIds(ids);
@@ -94,7 +94,7 @@ public class SmsSeckillPromotionController extends BaseController {
 
     @ApiOperation(value = "修改秒杀活动状态", httpMethod = "PUT")
     @ApiImplicitParams({
-            @ApiImplicitParam(name = "id", value = "秒杀活动id", required = true, paramType = "path", dataType = "Long"),
+            @ApiImplicitParam(name = "id", value = "秒杀活动id", required = true, paramType = "path", dataType = "Integer"),
             @ApiImplicitParam(name = "status", value = "显示状态", required = true, paramType = "path", dataType = "Integer")
     })
     @PutMapping("/id/{id}/status/{status}")
